Move importer and correlator into worker directory.
