+++ lsif/src/test-utils.ts
@@ -5,7 +5,7 @@ import { Connection } from 'typeorm'
 import { connectPostgres } from './connection'
 import { lsp } from 'lsif-protocol'
 import { child_process } from 'mz'
-import { convertLsif } from './importer'
+import { convertLsif } from './worker/importer/importer'
 import { XrepoDatabase } from './xrepo'
 import { dbFilename, ensureDirectory } from './util'
 import { userInfo } from 'os'
