+++ lsif/src/worker/processors/convert.ts
@@ -2,7 +2,7 @@ import * as constants from '../../constants'
 import * as fs from 'mz/fs'
 import * as path from 'path'
 import { ConfigurationFetcher } from '../../config'
-import { convertLsif } from '../../importer'
+import { convertLsif } from '../importer/importer'
 import { dbFilename } from '../../util'
 import { DBS_DIR_MAXIMUM_SIZE_BYTES, STORAGE_ROOT } from '../settings'
 import { logAndTraceCall, TracingContext } from '../../tracing'
